# Solve Problem 42 and Enhance Solving Problems from 29 to 32

+++ src/_3_problems_from_20_to_55/_3_1_problems_from_20_to_32/_3_1_11_problem_30/Problem30.cpp
@@ -192,7 +192,7 @@ void nextYear(
 void nextDecade(
     Date& date
 ) {
-    for (int year = 0; year < 10; ++year)
+    for (short year = 0; year < 10; ++year)
         nextYear(
             date
         );
